+++ src/resolvers/index.ts
@@ -1,0 +1,10 @@
+import { allTodos, addTodo } from './todos'
+
+export const resolvers = {
+  Query: {
+    allTodos,
+  },
+  Mutation: {
+    addTodo,
+  },
+}
