+++ src/typeDefs/mutation.graphql
@@ -1,0 +1,5 @@
+# import Todo from "todo.graphql"
+
+type Mutation {
+  addTodo(title: String!): Todo!
+}
