+++ src/typeDefs/query.graphql
@@ -1,0 +1,5 @@
+# import Todo from "todo.graphql"
+
+type Query {
+  allTodos: [Todo!]!
+}
