+++ src/typeDefs/schema.graphql
@@ -1,0 +1,7 @@
+# import Query from "query.graphql"
+# import Mutation from "mutation.graphql"
+
+schema {
+  query: Query
+  mutation: Mutation
+}
