+++ src/typeDefs/todo.graphql
@@ -1,0 +1,5 @@
+type Todo {
+  id: ID!
+  title: String!
+  done: Boolean!
+}
